modules/Single/Ollama/requirements.txt
kafka-python==2.0.2
requests==2.28.1
python-dotenv==1.0.0
nltk==3.7
numpy==1.23.1
